atividadesAula5/src/main/java/romanos/NumerosRomanos.java
package romanos;

import java.util.HashMap;
import java.util.Map;

public class NumerosRomanos {

    private static Map<Character, Integer> map;

    static {
        map = new HashMap<>();
        map.put('I', 1);
        map.put('V', 5);
        map.put('X', 10);
        map.put('L', 50);
        map.put('C', 100);
        map.put('D', 500);
        map.put('M', 1000);
    }

    public static int convert(String s) {
        int convertedNumber = 0;

        for (int i = 0; i < s.length(); i++) {
            int currentNumber = map.get(s.charAt(i));
            int next = i + 1 < s.length() ? map.get(s.charAt(i + 1)) : 0;

            if (currentNumber >= next) {
                convertedNumber += currentNumber;
            } else {
                convertedNumber -= currentNumber;
            }
        }

        return convertedNumber;
    }

    public static void main(String[] args){
        System.out.println(NumerosRomanos.traduzirNumeralRomano("MMID"));
    }



    //-1 indica retorno inválido
    public static int converteNumerosRomanos(String s){
        if (s==null || s.length()==0 || s.length() >9)
            return -1;



        return 0;
    }


    private static int traduzirNumeralRomano(String texto) {
        int n = 0;
        int numeralDaDireita = 0;
        for (int i = texto.length() - 1; i >= 0; i--) {
            int valor = (int) traduzirNumeralRomano(texto.charAt(i));
            n += valor * Math.signum(valor + 0.5 - numeralDaDireita);
            numeralDaDireita = valor;
        }
        return n;
    }

    private static double traduzirNumeralRomano(char caractere) {
        return Math.floor(Math.pow(10, "IXCM".indexOf(caractere))) + 5 * Math.floor(Math.pow(10, "VLD".indexOf(caractere)));
    }


}
